refactor: remove md helper and markdown-table dependency
